Issue: Fix cache invalidation problems
