+++ app/actions/createAccount/index.ts
@@ -1,6 +1,7 @@
 "use server";
 import prisma from "@lib/prisma";
 import { IsNotEmpty, IsString, MaxLength, validate } from "class-validator";
+import { revalidatePath } from "next/cache";
 import { isRedirectError } from "next/dist/client/components/redirect";
 import { redirect } from "next/navigation";
 
@@ -48,6 +49,7 @@ export default async function createAccount(_state: any, data: FormData) {
         },
       })
     ).id;
+    revalidatePath("/", "layout");
     redirect(`/user/${id}`);
   } catch (e) {
     if (isRedirectError(e)) {
